move plugins dependency class
