+++ src/main/kotlin/xyz/alexcrea/cuanvil/dependency/plugins/EcoEnchantDependency.kt
@@ -1,4 +1,4 @@
-package xyz.alexcrea.cuanvil.dependency
+package xyz.alexcrea.cuanvil.dependency.plugins
 
 import com.willfp.ecoenchants.enchant.EcoEnchant
 import com.willfp.ecoenchants.enchant.EcoEnchants
