+++ src/main/kotlin/xyz/alexcrea/cuanvil/dependency/plugins/LegacyEcoEnchantDependency.kt
@@ -1,4 +1,4 @@
-package xyz.alexcrea.cuanvil.dependency
+package xyz.alexcrea.cuanvil.dependency.plugins
 
 import com.willfp.ecoenchants.enchantments.EcoEnchant
 import com.willfp.ecoenchants.enchantments.EcoEnchants
